+++ src/Maw.Data.Abstractions/IVideoRepository.cs
@@ -13,13 +13,13 @@ public interface IVideoRepository
     Task<Category?> GetCategoryAsync(short categoryId, string[]? roles);
     Task<IEnumerable<Comment>> GetCommentsAsync(short videoId, string[] roles);
     Task<Rating?> GetRatingsAsync(short videoId, string username, string[] roles);
-    Task<GpsDetail?> GetGpsDetailAsync(int videoId, string[] roles);
+    Task<GpsDetail?> GetGpsDetailAsync(short videoId, string[] roles);
     Task InsertCommentAsync(short videoId, string username, string comment, string[] roles);
     Task<float?> SaveRatingAsync(short videoId, string username, short rating, string[] roles);
     Task<float?> RemoveRatingAsync(short videoId, string username, string[] roles);
 
     // admin functions
-    Task SetGpsOverrideAsync(int videoId, GpsCoordinate gps, string username);
-    Task<long> SetCategoryTeaserAsync(short categoryId, int videoId);
+    Task SetGpsOverrideAsync(short videoId, GpsCoordinate gps, string username);
+    Task<long> SetCategoryTeaserAsync(short categoryId, short videoId);
     Task<IEnumerable<CategoryAndRoles>> GetCategoriesAndRolesAsync();
 }
